Change cleanTarget to return a bool indicating success
Change `cleanTarget` and `cleanTargetLink`/`File` to return bool if they managed to create a clean filepath for the new link to be created at. This still needs to be implemented in `makeLink`, but the intent is to remove runtime errors and instead tell the user the link was unable to be created and move on. With this current implementation, it will be much easier to add a config option that causes the process to stop on the first fail.